requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.0.6
    # via kombu
apipkg==1.5
    # via execnet
asgiref==3.3.4
    # via django
attrs==20.3.0
    # via
    #   jsonschema
    #   pytest
billiard==3.6.4.0
    # via celery
boto3==1.17.34
    # via django-amazon-ses
botocore==1.20.34
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.1
    # via premailer
celery==5.1.2
    # via -r requirements/requirements.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
coverage==5.5
    # via pytest-cov
cssselect==1.1.0
    # via premailer
cssutils==2.2.0
    # via premailer
dj-database-url==0.5.0
    # via -r requirements/requirements.in
django==3.2.6
    # via
    #   -r requirements/requirements.in
    #   django-amazon-ses
    #   django-cors-headers
    #   django-storages
    #   django-watchman
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   model-bakery
django-amazon-ses==4.0.0
    # via -r requirements/requirements.in
django-ckeditor==6.1.0
    # via -r requirements/requirements.in
django-cleanup==5.2.0
    # via -r requirements/requirements.in
django-cors-headers==3.8.0
    # via -r requirements/requirements.in
django-js-asset==1.2.2
    # via django-ckeditor
django-solo==1.1.5
    # via -r requirements/requirements.in
django-storages==1.11.1
    # via -r requirements/requirements.in
django-tabbed-admin @ git+https://github.com/CornerCaseTechnologies/django-tabbed-admin.git
    # via -r requirements/requirements.in
django-watchman==1.2.0
    # via -r requirements/requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==4.7.2
    # via -r requirements/requirements.in
drf-spectacular==0.17.3
    # via -r requirements/requirements.in
execnet==1.8.0
    # via pytest-xdist
gunicorn==20.1.0
    # via -r requirements/requirements.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via drf-spectacular
kombu==5.1.0
    # via celery
lxml==4.6.3
    # via premailer
model-bakery==1.3.2
    # via -r requirements/requirements.in
packaging==20.9
    # via pytest
parameterized==0.8.1
    # via -r requirements/requirements.in
pillow==8.2.0
    # via -r requirements/requirements.in
pluggy==0.13.1
    # via pytest
premailer==3.8.0
    # via -r requirements/requirements.in
prompt-toolkit==3.0.18
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements/requirements.in
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pyjwt==2.0.1
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pyrsistent==0.17.3
    # via jsonschema
pytest==6.2.2
    # via
    #   -r requirements/requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-xdist
pytest-cov==2.11.1
    # via -r requirements/requirements.in
pytest-django==4.1.0
    # via -r requirements/requirements.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-xdist==2.2.1
    # via -r requirements/requirements.in
python-dateutil==2.8.1
    # via botocore
pytz==2021.1
    # via
    #   celery
    #   django
pyyaml==5.4.1
    # via drf-spectacular
redis==3.5.3
    # via -r requirements/requirements.in
requests==2.25.1
    # via
    #   premailer
    #   responses
responses==0.13.1
    # via -r requirements/requirements.in
s3transfer==0.3.6
    # via boto3
sentry-sdk==1.0.0
    # via -r requirements/requirements.in
six==1.15.0
    # via
    #   click-repl
    #   jsonschema
    #   python-dateutil
    #   responses
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via pytest
typing-extensions==3.7.4.3
    # via -r requirements/requirements.in
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.4
    # via
    #   botocore
    #   requests
    #   responses
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
